#884 [RENAME] find-id-pw-fail.svg 파일명 변경 -> taskFailedIllustration.svg
