+++ src/page/account/NotFoundIdPage/NotFoundIdPage.jsx
@@ -5,7 +5,7 @@ import { BackAppBar, Button } from '@/shared/component';
 
 import styles from './NotFoundIdPage.module.css';
 
-import findIdPwFail from '@/assets/images/find-id-pw-fail.svg';
+import findIdPwFail from '@/assets/images/taskFailedIllustration.svg';
 
 export default function NotFoundIdPage() {
   const navigate = useNavigate();
